requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
python-docx
pillow
python-multipart